Move name class into name package in model package from person package.
