+++ src/main/java/seedu/address/model/group/UniqueGroupList.java
@@ -1,0 +1,4 @@
+package seedu.address.model.group;
+
+public class UniqueGroupList {
+}
