+++ src/main/java/seedu/address/model/names/Name.java
@@ -1,10 +1,10 @@
-package seedu.address.model.person;
+package seedu.address.model.names;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
 /**
- * Represents a Person's name in the address book.
+ * Represents a Name in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidName(String)}
  */
 public class Name {
