+++ src/test/java/seedu/address/storage/JsonAdaptedPersonTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
+import seedu.address.model.names.Name;
 import seedu.address.model.person.Phone;
 
 public class JsonAdaptedPersonTest {
